Various updates
Rebased underlying CNN feature extractor with newer Caffe version
and fixes to make the code compile and work on Fedora 29.

Updated README with the final journal reference and link to
pre-print version of the paper.

Fixed the Caffe library path in startup script (the currently
used version installs under lib64 instead of lib).